refactor  Code
