+++ untitled/.gitignore
@@ -5,7 +5,6 @@ build/
 !**/src/test/**/build/
 
 ### IntelliJ IDEA ###
-.idea
 .idea/modules.xml
 .idea/jarRepositories.xml
 .idea/compiler.xml
