+++ untitled/src/main/java/com/sprint/mission/Main.java
@@ -1,0 +1,7 @@
+package com.sprint.mission;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello, World!");
+    }
+}
